+++ src/test/groovy/de/dfki/mary/voicebuilding/LegacyPluginFunctionalTest.groovy
@@ -1,9 +1,12 @@
 package de.dfki.mary.voicebuilding
 
 import org.gradle.testkit.runner.GradleRunner
-import org.testng.annotations.*
+import org.testng.annotations.BeforeSuite
+import org.testng.annotations.DataProvider
+import org.testng.annotations.Test
 
-import static org.gradle.testkit.runner.TaskOutcome.*
+import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
+import static org.gradle.testkit.runner.TaskOutcome.UP_TO_DATE
 
 class LegacyPluginFunctionalTest {
 
@@ -13,7 +16,7 @@ class LegacyPluginFunctionalTest {
     void setup() {
         def projectDir = File.createTempDir()
 
-        gradle = GradleRunner.create().withProjectDir(projectDir).withPluginClasspath()
+        gradle = GradleRunner.create().withProjectDir(projectDir).withPluginClasspath().forwardOutput()
 
         // Add the logic under test to the test build
         new File(projectDir, 'gradle.properties').withWriter {
@@ -77,13 +80,11 @@ class LegacyPluginFunctionalTest {
 
     @Test(dataProvider = 'taskNames')
     void testTasks(String taskName, boolean runTestTask) {
-        def result = gradle.withArguments(taskName).build()
-        println result.output
+        def result = gradle.withArguments(taskName, '-s').build()
         assert result.task(":$taskName").outcome in [SUCCESS, UP_TO_DATE]
         if (runTestTask) {
             def testTaskName = 'test' + taskName.capitalize()
             result = gradle.withArguments(testTaskName).build()
-            println result.output
             assert result.task(":$taskName").outcome == UP_TO_DATE
             assert result.task(":$testTaskName").outcome == SUCCESS
         }